Move tabs new a new component file
